third_party/blink/web_tests/svg/animations/animate-elem-08-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getTransformToElement(rootSVGElement, path1).a is 0
PASS getTransformToElement(rootSVGElement, path1).b is 1
PASS getTransformToElement(rootSVGElement, path1).c is -1
PASS getTransformToElement(rootSVGElement, path1).d is 0
PASS getTransformToElement(rootSVGElement, path1).e is 224.9
PASS getTransformToElement(rootSVGElement, path1).f is -25
PASS getTransformToElement(rootSVGElement, path2).a is 0
PASS getTransformToElement(rootSVGElement, path2).b is -1
PASS getTransformToElement(rootSVGElement, path2).c is 1
PASS getTransformToElement(rootSVGElement, path2).d is 0
PASS getTransformToElement(rootSVGElement, path2).e is -225
PASS getTransformToElement(rootSVGElement, path2).f is 275
PASS getTransformToElement(rootSVGElement, path1).a is 1
PASS getTransformToElement(rootSVGElement, path1).b is 0.1
PASS getTransformToElement(rootSVGElement, path1).c is -0.1
PASS getTransformToElement(rootSVGElement, path1).d is 1
PASS getTransformToElement(rootSVGElement, path1).e is -70.9
PASS getTransformToElement(rootSVGElement, path1).f is -182.6
PASS getTransformToElement(rootSVGElement, path2).a is -1
PASS getTransformToElement(rootSVGElement, path2).b is -0.1
PASS getTransformToElement(rootSVGElement, path2).c is 0.1
PASS getTransformToElement(rootSVGElement, path2).d is -1
PASS getTransformToElement(rootSVGElement, path2).e is 319.3
PASS getTransformToElement(rootSVGElement, path2).f is 210.7
PASS getTransformToElement(rootSVGElement, path1).a is 0.7
PASS getTransformToElement(rootSVGElement, path1).b is -0.7
PASS getTransformToElement(rootSVGElement, path1).c is 0.7
PASS getTransformToElement(rootSVGElement, path1).d is 0.7
PASS getTransformToElement(rootSVGElement, path1).e is -265.1
PASS getTransformToElement(rootSVGElement, path1).f is -17.7
PASS getTransformToElement(rootSVGElement, path2).a is -0.7
PASS getTransformToElement(rootSVGElement, path2).b is 0.7
PASS getTransformToElement(rootSVGElement, path2).c is -0.7
PASS getTransformToElement(rootSVGElement, path2).d is -0.7
PASS getTransformToElement(rootSVGElement, path2).e is 442
PASS getTransformToElement(rootSVGElement, path2).f is -159
PASS getTransformToElement(rootSVGElement, path1).a is 0.7
PASS getTransformToElement(rootSVGElement, path1).b is -0.7
PASS getTransformToElement(rootSVGElement, path1).c is 0.7
PASS getTransformToElement(rootSVGElement, path1).d is 0.7
PASS getTransformToElement(rootSVGElement, path1).e is -265.1
PASS getTransformToElement(rootSVGElement, path1).f is -17.7
PASS getTransformToElement(rootSVGElement, path2).a is -0.7
PASS getTransformToElement(rootSVGElement, path2).b is 0.7
PASS getTransformToElement(rootSVGElement, path2).c is -0.7
PASS getTransformToElement(rootSVGElement, path2).d is -0.7
PASS getTransformToElement(rootSVGElement, path2).e is 442
PASS getTransformToElement(rootSVGElement, path2).f is -159
PASS successfullyParsed is true

TEST COMPLETE
